AutoComplete Rework and various changes working in progress update and save
