--- src/main/resources/CLIGraphics.fxml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<VBox prefHeight="720.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/22"
-  xmlns:fx="http://javafx.com/fxml/1" style="-fx-background-color: black;">
-  <AnchorPane prefHeight="720.0" prefWidth="1280.0">
-    <Text layoutX="20.0" layoutY="60.0" fill="white"
-      text="Welcome to the Unified Modeling Language Editor" wrappingWidth="1240.0">
-      <font>
-        <Font name="Consolas" size="24.0"/>
-      </font>
-    </Text>
-    <Text layoutX="20.0" layoutY="100.0" fill="white" text="Please choose an interface to start:"
-      wrappingWidth="1240.0">
-      <font>
-        <Font name="Consolas" size="18.0"/>
-      </font>
-    </Text>
-    <Text layoutX="20.0" layoutY="140.0" fill="white" text="- GUI (Type 'gui' or 'g')"
-      wrappingWidth="1240.0">
-      <font>
-        <Font name="Consolas" size="16.0"/>
-      </font>
-    </Text>
-    <Text layoutX="20.0" layoutY="180.0" fill="white" text="- CLI (Type 'cli' or 'c')"
-      wrappingWidth="1240.0">
-      <font>
-        <Font name="Consolas" size="16.0"/>
-      </font>
-    </Text>
-    <Text layoutX="20.0" layoutY="220.0" fill="white" text="- Exit (Type 'exit' or 'quit')"
-      wrappingWidth="1240.0">
-      <font>
-        <Font name="Consolas" size="16.0"/>
-      </font>
-    </Text>
-    <Button layoutX="20.0" layoutY="260.0" mnemonicParsing="false" text="Start GUI"
-      prefWidth="200.0" fx:id="startGuiButton"/>
-    <Button layoutX="240.0" layoutY="260.0" mnemonicParsing="false" text="Start CLI"
-      prefWidth="200.0" fx:id="startCliButton"/>
-    <Button layoutX="460.0" layoutY="260.0" mnemonicParsing="false" text="Exit" prefWidth="200.0"
-      fx:id="exitButton"/>
-  </AnchorPane>
-</VBox>
