+++ src/main/resources/MainMenu.fxml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<VBox alignment="CENTER" prefHeight="330.0" prefWidth="900.0" spacing="30" style="-fx-background-color: black;" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.project.Controller.MainMenuController">
+  <Text fill="white" text="Welcome to the Unified Modeling Language Editor" textAlignment="CENTER" wrappingWidth="900.0">
+    <font>
+      <Font name="Consolas" size="28.0" />
+    </font>
+  </Text>
+
+  <Text fill="white" text="Please choose an interface to start:" textAlignment="CENTER" wrappingWidth="900.0">
+    <font>
+      <Font name="Consolas" size="24.0" />
+    </font>
+  </Text>
+
+  <HBox alignment="CENTER_LEFT" prefHeight="25.0" prefWidth="769.0" spacing="20" style="-fx-padding: 0 100 0 100;">
+    <Text fill="white" text="- Graphical User Interface" textAlignment="RIGHT" wrappingWidth="297.3333511352539">
+      <font>
+        <Font name="Consolas" size="20.0" />
+      </font>
+    </Text>
+    <Button fx:id="startGuiButton" mnemonicParsing="false" prefHeight="25.0" prefWidth="240.0" text="Start GUI" textAlignment="CENTER" onAction="#handleButtonAction" />
+  </HBox>
+
+  <HBox alignment="CENTER_LEFT" spacing="20" style="-fx-padding: 0 100 0 100;">
+    <Text fill="white" text="- Command Line Interface" textAlignment="RIGHT" wrappingWidth="298.66664123535156">
+      <font>
+        <Font name="Consolas" size="20.0" />
+      </font>
+    </Text>
+    <Button fx:id="startCliButton" mnemonicParsing="false" prefHeight="25.0" prefWidth="239.0" text="Start CLI" onAction="#handleButtonAction" />
+  </HBox>
+
+  <HBox alignment="CENTER_LEFT" spacing="20" style="-fx-padding: 0 100 0 100;">
+    <Text fill="white" text="- Terminate This Menu" textAlignment="RIGHT" wrappingWidth="297.9999771118164">
+      <font>
+        <Font name="Consolas" size="20.0" />
+      </font>
+    </Text>
+    <Button fx:id="exitButton" mnemonicParsing="false" prefHeight="25.0" prefWidth="239.0" text="Exit" textFill="RED" onAction="#handleButtonAction" />
+  </HBox>
+</VBox>
